+++ R/expr__expr.R
@@ -3420,3 +3420,105 @@ Expr_rle_id = function() {
   .pr$Expr$rle_id(self) |>
     unwrap("in $rle_id():")
 }
+
+#' Bin continuous values into discrete categories
+#'
+#' @param breaks Unique cut points.
+#' @param ... Ignored.
+#' @param labels Names of the categories. The number of labels must be equal to
+#' the number of cut points plus one.
+#' @param left_closed Set the intervals to be left-closed instead of right-closed.
+#' @param include_breaks Include a column with the right endpoint of the bin each
+#' observation falls in. This will change the data type of the output from a
+#' [`Categorical`][DataType_Categorical] to a [`Struct`][DataType_Struct].
+#'
+#' @return Expr of data type `Categorical` is `include_breaks` is `FALSE` and
+#' of data type `Struct` if `include_breaks` is `TRUE`.
+#'
+#' @seealso [`$qcut()`][Expr_qcut]
+#'
+#' @examples
+#' df = pl$DataFrame(foo = c(-2, -1, 0, 1, 2))
+#'
+#' df$with_columns(
+#'   cut = pl$col("foo")$cut(c(-1, 1), labels = c("a", "b", "c"))
+#' )
+#'
+#' # Add both the category and the breakpoint
+#' df$with_columns(
+#'   cut = pl$col("foo")$cut(c(-1, 1), include_breaks = TRUE)
+#' )$unnest("cut")
+Expr_cut = function(breaks, ..., labels = NULL, left_closed = FALSE, include_breaks = FALSE) {
+  .pr$Expr$cut(
+    self,
+    breaks = breaks,
+    labels = labels,
+    left_closed = left_closed,
+    include_breaks = include_breaks
+  ) |>
+    unwrap("in $cut():")
+}
+
+#' Bin continuous values into discrete categories based on their quantiles
+#'
+#' @param quantiles Either a vector of quantile probabilities between 0 and 1 or
+#' a positive integer determining the number of bins with uniform probability.
+#' @param allow_duplicates If set to `TRUE`, duplicates in the resulting
+#' quantiles are dropped, rather than raising an error. This can happen even
+#' with unique probabilities, depending on the data.
+#'
+#' @inherit Expr_cut params return
+#'
+#' @seealso [`$cut()`][Expr_cut]
+#'
+#' @examples
+#' df = pl$DataFrame(foo = c(-2, -1, 0, 1, 2))
+#'
+#' # Divide a column into three categories according to pre-defined quantile
+#' # probabilities
+#' df$with_columns(
+#'   qcut = pl$col("foo")$qcut(c(0.25, 0.75), labels = c("a", "b", "c"))
+#' )
+#'
+#' # Divide a column into two categories using uniform quantile probabilities.
+#' df$with_columns(
+#'   qcut = pl$col("foo")$qcut(2, labels = c("low", "high"), left_closed = TRUE)
+#' )
+#'
+#' # Add both the category and the breakpoint
+#' df$with_columns(
+#'   qcut = pl$col("foo")$qcut(c(0.25, 0.75), include_breaks = TRUE)
+#' )$unnest("qcut")
+Expr_qcut = function(
+    quantiles,
+    ...,
+    labels = NULL,
+    left_closed = FALSE,
+    allow_duplicates = FALSE,
+    include_breaks = FALSE) {
+  if (length(quantiles) == 1) {
+    if (!is.numeric(quantiles) || as.integer(quantiles) != quantiles) {
+      Err_plain("`quantiles` must either be an integer of length 1 or a vector of probabilities.") |>
+        unwrap("in $qcut():")
+    }
+    .pr$Expr$qcut_uniform(
+      self,
+      n_bins = quantiles,
+      labels = labels,
+      left_closed = left_closed,
+      allow_duplicates = allow_duplicates,
+      include_breaks = include_breaks
+    ) |>
+      unwrap("in $qcut():")
+  } else {
+    .pr$Expr$qcut(
+      self,
+      probs = quantiles,
+      labels = labels,
+      left_closed = left_closed,
+      allow_duplicates = allow_duplicates,
+      include_breaks = include_breaks
+    ) |>
+      unwrap("in $qcut():")
+  }
+}