+++ NEWS.md
@@ -2,6 +2,15 @@
 
 ## polars (development version)
 
+### New features
+
+- `$cut()` and `$qcut()` to bin continuous values into discrete categories (#1057).
+
+### Bug fixes
+
+- In some read/scan functions, downloading files could fail if the URL was too
+  long. This is now fixed (#1049, @DyfanJones).
+
 ## polars 0.16.1
 
 This is a small hot-fix release to update dependent Rust polars to 0.39.1 (#1042).
@@ -10,9 +19,7 @@ Also, there are some updates.
 
 ### Bug fixes
 
-* `$len()` now correctly includes `null` values in the count (#1044).
-* In some read/scan functions, downloading files could fail if the URL was too
-  long. This is now fixed (#1049, @DyfanJones).
+- `$len()` now correctly includes `null` values in the count (#1044).
 
 ### Other improvements